[DEVOPS-834] Fix faucet static serving
